\name{loadAPI-package}
\alias{loadAPI-package}
\alias{loadAPI}
\docType{package}
\title{
\packageTitle{loadAPI}
}
\description{
\packageDescription{loadAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{loadAPI}
\packageIndices{loadAPI}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{loadAPI}

Maintainer: \packageMaintainer{loadAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
