% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadAPI.R
\name{geocode_GET}
\alias{geocode_GET}
\title{Create GET request}
\usage{
geocode_GET(address)
}
\arguments{
\item{address}{the address is real address that you wish to search.}
}
\description{
Make GET request to Google Geocode API.
}

